refactor(run-types): move the credit-card pure fns to their own module
